requirements.txt
Flask==2.3.2
gunicorn==20.1.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
webdriver-manager==4.1.0
apscheduler==3.10.1
python-dotenv==1.0.1
pytest==7.4.0
python-json-logger==2.0.7
openai
